+++ src/pages/dashboard/Renderer/Renderer/Main.tsx
@@ -3,6 +3,7 @@ import _ from 'lodash';
 import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
 import { Dropdown, Menu, Tooltip, Space, Drawer, message } from 'antd';
+import type { MenuProps } from 'antd';
 import {
   InfoCircleOutlined,
   MoreOutlined,
@@ -22,6 +23,8 @@ import {
 
 import { IRawTimeRange } from '@/components/TimeRangePicker';
 import replaceTemplateVariables from '@/pages/dashboard/Variables/utils/replaceTemplateVariables';
+import type { DashboardQueryState } from '../datasource/types';
+import type { CalculatedSeries } from '../utils/getCalculatedValuesBySeries';
 
 import PanelEmpty from '../components/PanelEmpty';
 import CloneIcon from '../components/CloneIcon';
@@ -45,14 +48,7 @@ import { IProps } from './index';
 function index(
   props: IProps & {
     controllersVisible: boolean;
-    queryResult: {
-      query: any[];
-      series: any[];
-      error: string;
-      loading: boolean;
-      loaded: boolean;
-      range: IRawTimeRange;
-    };
+    queryResult: DashboardQueryState;
     containerEleRef: React.RefObject<HTMLDivElement>;
     time: IRawTimeRange;
     setTime?: (time: IRawTimeRange) => void;
@@ -88,10 +84,10 @@ function index(
   } = props;
   const [visible, setVisible] = useState(false);
   const values = _.cloneDeep(props.values);
-  const tableRef = useRef<any>(null);
-  const tableNGRef = useRef<any>(null);
+  const tableRef = useRef<{ exportCsv: () => void }>(null);
+  const tableNGRef = useRef<{ exportCsv: () => void }>(null);
   const bodyWrapRef = useRef<HTMLDivElement>(null);
-  const { query, series, error, loading, loaded, range } = queryResult;
+  const { query, series, error, loading, loaded, range, revision } = queryResult;
   const name = replaceTemplateVariables(values.name, {
     scopedVars: values.scopedVars,
     range: time,
@@ -105,14 +101,188 @@ function index(
 
   // TODO: 如果 hexbin 的 colorRange 为 string 时转成成 array
   if (typeof _.get(values, 'custom.colorRange') === 'string') {
-    _.set(values, 'custom.colorRange', _.split(_.get(values, 'custom.colorRange'), ','));
+    _.set(values, 'custom.colorRange', _.split(_.get(values, 'custom.colorRange') as string, ','));
   }
   const subProps = {
     id,
     values,
-    series,
+    // DashboardSeries 含日志序列（metric 为 unknown），渲染器按 CalculatedSeries 消费，运行时结构兼容
+    series: series as CalculatedSeries[],
+    dataRevision: revision,
     onOverridesChange,
   };
+  const menuItems: MenuProps['items'] = [
+    {
+      key: 'review_btn',
+      label: (
+        <Space>
+          <EyeOutlined />
+          {t('common:btn.view')}
+        </Space>
+      ),
+      onClick: () => {
+        setVisible(false);
+        setViewModalVisible(true);
+      },
+    },
+    {
+      key: 'refresh_btn',
+      label: (
+        <Space>
+          <SyncOutlined />
+          {t('refresh_btn')}
+        </Space>
+      ),
+      onClick: () => {
+        setVisible(true);
+        setTime?.({ ...time, refreshFlag: _.uniqueId('refreshFlag_ ') });
+      },
+    },
+    ...(isAuthorized && !values.repeatPanelId
+      ? [
+          {
+            key: 'edit_btn',
+            label: (
+              <Space>
+                <SettingOutlined />
+                {t('common:btn.edit')}
+              </Space>
+            ),
+            onClick: () => {
+              setVisible(false);
+              onEditClick?.(panelWidth);
+            },
+          },
+          {
+            key: 'clone_btn',
+            label: (
+              <Space>
+                <CloneIcon />
+                {t('common:btn.clone')}
+              </Space>
+            ),
+            onClick: () => {
+              setVisible(false);
+              onCloneClick?.();
+            },
+          },
+          {
+            key: 'copy_btn',
+            label: (
+              <Space>
+                <CopyOutlined />
+                {t('common:btn.copy')}
+              </Space>
+            ),
+            onClick: () => {
+              setVisible(false);
+              void onCopyClick?.();
+            },
+          },
+        ]
+      : []),
+    {
+      key: 'share_btn',
+      label: (
+        <Space>
+          <ShareAltOutlined />
+          {t('share_btn')}
+        </Space>
+      ),
+      onClick: () => {
+        setVisible(false);
+        onShareClick?.();
+      },
+    },
+    ...(values.type === 'table'
+      ? [
+          {
+            key: 'export_btn',
+            label: (
+              <Space>
+                <ExportOutlined />
+                {t('export_btn')}
+              </Space>
+            ),
+            onClick: () => {
+              tableRef.current?.exportCsv();
+              setVisible(false);
+            },
+          },
+        ]
+      : []),
+    ...(values.type === 'tableNG'
+      ? [
+          {
+            key: 'export_btn',
+            label: (
+              <Space>
+                <ExportOutlined />
+                {t('export_btn')}
+              </Space>
+            ),
+            onClick: () => {
+              tableNGRef.current?.exportCsv();
+              setVisible(false);
+            },
+          },
+        ]
+      : []),
+    ...(values.type === 'table'
+      ? [
+          {
+            key: 'clear_cache_btn',
+            label: (
+              <Tooltip title={t('clear_cache_btn_tip')} placement='left'>
+                <Space>
+                  <ClearOutlined />
+                  {t('clear_cache_btn')}
+                </Space>
+              </Tooltip>
+            ),
+            onClick: () => {
+              window.localStorage.removeItem(`dashboard-table2.1-resizable-${values.id}`);
+              setVisible(false);
+            },
+          },
+        ]
+      : []),
+    ...(!isPreview
+      ? [
+          {
+            key: 'inspect_btn',
+            label: (
+              <Space>
+                <InfoCircleOutlined />
+                {t('inspect_btn')}
+              </Space>
+            ),
+            onClick: () => {
+              setVisible(false);
+              setTime?.({ ...time, refreshFlag: _.uniqueId('refreshFlag_ ') });
+              setInspect(true);
+            },
+          },
+        ]
+      : []),
+    ...(isAuthorized && !values.repeatPanelId
+      ? [
+          {
+            key: 'delete_btn',
+            label: (
+              <Space>
+                <DeleteOutlined />
+                {t('common:btn.delete')}
+              </Space>
+            ),
+            onClick: () => {
+              setVisible(false);
+              onDeleteClick?.();
+            },
+          },
+        ]
+      : []),
+  ];
 
   const RendererCptMap = {
     timeseries: () => (
@@ -129,7 +299,8 @@ function index(
     ),
     stat: () => <Stat {...subProps} bodyWrapRef={bodyWrapRef} themeMode={themeMode} isPreview={isPreview} />,
     table: () => <Table {...subProps} themeMode={themeMode} isPreview={isPreview} ref={tableRef} />,
-    tableNG: () => <TableNG {...subProps} themeMode={themeMode} isPreview={isPreview} ref={tableNGRef} />,
+    // TableNG 的 series 类型为 DashboardSeries[]，与其他渲染器（CalculatedSeries[]）不同
+    tableNG: () => <TableNG {...subProps} series={series} themeMode={themeMode} isPreview={isPreview} ref={tableNGRef} />,
     pie: () => <Pie {...subProps} themeMode={themeMode} isPreview={isPreview} />,
     hexbin: () => <Hexbin {...subProps} themeMode={themeMode} isPreview={isPreview} />,
     barGauge: () => <BarGauge {...subProps} themeMode={themeMode} isPreview={isPreview} />,
@@ -243,172 +414,7 @@ function index(
                     onVisibleChange={(visible) => {
                       setVisible(visible);
                     }}
-                    overlay={
-                      <Menu>
-                        <Menu.Item
-                          onClick={() => {
-                            setVisible(false);
-                            setViewModalVisible(true);
-                          }}
-                          key='review_btn'
-                        >
-                          <Space>
-                            <EyeOutlined />
-                            {t('common:btn.view')}
-                          </Space>
-                        </Menu.Item>
-
-                        <Menu.Item
-                          onClick={() => {
-                            setVisible(true);
-                            setTime &&
-                              setTime({
-                                ...time,
-                                refreshFlag: _.uniqueId('refreshFlag_ '),
-                              });
-                          }}
-                          key='refresh_btn'
-                        >
-                          <Space>
-                            <SyncOutlined />
-                            {t('refresh_btn')}
-                          </Space>
-                        </Menu.Item>
-                        {isAuthorized && !values.repeatPanelId && (
-                          <Menu.Item
-                            onClick={() => {
-                              setVisible(false);
-                              if (onEditClick) onEditClick(panelWidth);
-                            }}
-                            key='edit_btn'
-                          >
-                            <Space>
-                              <SettingOutlined />
-                              {t('common:btn.edit')}
-                            </Space>
-                          </Menu.Item>
-                        )}
-                        {isAuthorized && !values.repeatPanelId && (
-                          <Menu.Item
-                            onClick={() => {
-                              setVisible(false);
-                              if (onCloneClick) onCloneClick();
-                            }}
-                            key='clone_btn'
-                          >
-                            <Space>
-                              <CloneIcon />
-                              {t('common:btn.clone')}
-                            </Space>
-                          </Menu.Item>
-                        )}
-                        {isAuthorized && !values.repeatPanelId && (
-                          <Menu.Item
-                            onClick={() => {
-                              setVisible(false);
-                              if (onCopyClick) {
-                                void onCopyClick();
-                              }
-                            }}
-                            key='copy_btn'
-                          >
-                            <Space>
-                              <CopyOutlined />
-                              {t('common:btn.copy')}
-                            </Space>
-                          </Menu.Item>
-                        )}
-                        <Menu.Item
-                          onClick={() => {
-                            setVisible(false);
-                            if (onShareClick) onShareClick();
-                          }}
-                          key='share_btn'
-                        >
-                          <Space>
-                            <ShareAltOutlined />
-                            {t('share_btn')}
-                          </Space>
-                        </Menu.Item>
-                        {values.type === 'table' && (
-                          <Menu.Item
-                            onClick={() => {
-                              tableRef.current.exportCsv();
-                              setVisible(false);
-                            }}
-                            key='export_btn'
-                          >
-                            <Space>
-                              <ExportOutlined />
-                              {t('export_btn')}
-                            </Space>
-                          </Menu.Item>
-                        )}
-                        {values.type === 'tableNG' && (
-                          <Menu.Item
-                            onClick={() => {
-                              tableNGRef.current.exportCsv();
-                              setVisible(false);
-                            }}
-                            key='export_btn'
-                          >
-                            <Space>
-                              <ExportOutlined />
-                              {t('export_btn')}
-                            </Space>
-                          </Menu.Item>
-                        )}
-                        {values.type === 'table' && (
-                          <Tooltip title={t('clear_cache_btn_tip')} placement='left'>
-                            <Menu.Item
-                              onClick={() => {
-                                window.localStorage.removeItem(`dashboard-table2.1-resizable-${values.id}`);
-                                setVisible(false);
-                              }}
-                              key='clear_cache_btn'
-                            >
-                              <Space>
-                                <ClearOutlined />
-                                {t('clear_cache_btn')}
-                              </Space>
-                            </Menu.Item>
-                          </Tooltip>
-                        )}
-                        {!isPreview && (
-                          <Menu.Item
-                            onClick={() => {
-                              setVisible(false);
-                              setTime &&
-                                setTime({
-                                  ...time,
-                                  refreshFlag: _.uniqueId('refreshFlag_ '),
-                                });
-                              setInspect(true);
-                            }}
-                            key='inspect_btn'
-                          >
-                            <Space>
-                              <InfoCircleOutlined />
-                              {t('inspect_btn')}
-                            </Space>
-                          </Menu.Item>
-                        )}
-                        {isAuthorized && !values.repeatPanelId && (
-                          <Menu.Item
-                            onClick={() => {
-                              setVisible(false);
-                              if (onDeleteClick) onDeleteClick();
-                            }}
-                            key='delete_btn'
-                          >
-                            <Space>
-                              <DeleteOutlined />
-                              {t('common:btn.delete')}
-                            </Space>
-                          </Menu.Item>
-                        )}
-                      </Menu>
-                    }
+                    overlay={<Menu items={menuItems} />}
                   >
                     <MoreOutlined className='renderer-header-controller' />
                   </Dropdown>